pyproject.toml
[project]
name = "custom-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.84.0",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]
